renamed home/Register -> home/CTA
