+++ src/pages/home/index.js
@@ -1,17 +1,17 @@
 import Hero from './Hero';
 import Features from './Features';
-import Register from './Register';
+import CTA from './CTA';
 import Members from './Members';
 import { Footer } from '../../components';
 
 export default function App() {
 	return (
 		<>
-			<Hero></Hero>
-			<Features></Features>
-			<Register></Register>
-			<Members></Members>
-			<Footer></Footer>
+			<Hero />
+			<Features />
+			<CTA />
+			<Members />
+			<Footer />
 		</>
 	);
 }
